Strip "Host" header when forwarding too
When proxying through Squid, the proxy replaces the Host header with the
right one for the destination URL. But httpx by itself apparently
assumes that if we manually set the Host header we must have wanted it
that way. Which is an absolutely fair assumption but is not true here.