Also filter Content-Encoding response header
httpx is decoding whatever Content-Encoding the server returns for us,
so what we return to the client is no longer in that encoding. Maybe it
should be, but that's a separate issue.

That said, we can't remove the header from the copy that httpx is
looking at to decode the response, so we need a separate copy of the
response headers to clean up.