Use allow-list for request headers, block-list for reply
I keep having to add more request headers to the block-list which seems
like a losing battle. Switch to denying all but a few allowed request
headers instead.

Response headers seem like they're usually okay if they weren't prompted
by some special request header, so for now I'm sticking with the
block-list approach on responses. We'll see if that turns out to be a
mistake.